Cracking the Coding/requirements.txt.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
docopt==0.6.2
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
kafka==1.3.5
kafka-python==2.0.2
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pipreqs==0.4.10
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
yarg==0.1.9
